single responsibility principle for getCroppedImage.
