+++ lib/cropImage/getRadianAngle.ts
@@ -1,0 +1,3 @@
+export function getRadianAngle(degreeValue: number): number {
+  return (degreeValue * Math.PI) / 180;
+}
